Blueprint refactoring.
