+++ app/templates/_temp.html
@@ -12,27 +12,30 @@
 <br />
 
 <h3>Forgot password</h3>
-<a href="forgot-password" target="_blank">Forgot password</a> =>
-<a href="forgot-password?error=failed" target="_blank">Invalid email</a> =>
-<a href="forgot-password/check-email/" target="_blank">Check email</a>
+<a href="passwords/forgot-password" target="_blank">Forgot password</a> =>
+<a href="passwords/forgot-password?error=failed" target="_blank">Invalid email</a> =>
+<a href="passwords/forgot-password/check-email/" target="_blank">Check email</a>
 <br />
 <br />
 
 <h3>Reset password</h3>
-<a href="reset-password" target="_blank">Reset password</a><br />
-<a href="reset-password?error=password-mismatch" target="_blank">Password mismatch</a><br />
-<a href="reset-password?error=stronger-password-required" target="_blank">Password not strong enough</a><br />
-<a href="reset-password?error=password-too-common" target="_blank">Password too common</a><br />
-<a href="reset-password/confirmation" target="_blank">Confirmation</a><br />
+<a href="passwords/reset-password" target="_blank">Reset password</a><br />
+<a href="passwords/reset-password?error=password-mismatch" target="_blank">Password mismatch</a><br />
+<a href="passwords/reset-password?error=stronger-password-required" target="_blank">Password not strong enough</a><br />
+<a href="passwords/reset-password?error=password-too-common" target="_blank">Password too common</a><br />
+<a href="passwords/reset-password/confirmation" target="_blank">Confirmation</a><br />
 <br />
 
 <h3>Register</h3>
-<a href="create-account" target="_blank">Create an account (enter activation code)</a><br />
-<a href="create-account/confirm-organisation-survey/" target="_blank">Confirm organisation and survey</a><br />
-<a href="create-account/enter-account-details/" target="_blank">Enter your account details</a><br />
-<a href="create-account/check-email/" target="_blank">Almost done...</a><br />
-<a href="create-account/activate-account" target="_blank">Activate your account</a><br />
+<a href="register/create-account" target="_blank">Create an account (enter activation code)</a><br />
+<a href="register/create-account/confirm-organisation-survey/" target="_blank">Confirm organisation and survey</a><br />
+<a href="register/create-account/enter-account-details/" target="_blank">Enter your account details</a><br />
+<a href="register/create-account/check-email/" target="_blank">Almost done...</a><br />
+<a href="register/create-account/activate-account" target="_blank">Activate your account</a><br />
 <br />
+
+<h3>Email re-sent</h3>
+<a href="register/create-account/email-resent?user_id=TODO" target="_blank">Email re-sent</a><br />
 
 <h3>Surveys</h3>
 <a href="/surveys" target="_blank">To Do</a><br />
@@ -41,4 +44,3 @@
 
 <h3>Secure Messages</h3>
 <a href="secure-messages/view" target="_blank">View secure message</a>
-
