--- app/templates/_tempLogedIn.html
@@ -1,29 +1,0 @@
-<br />
-<h1>Congratulations You are Logged In</h1>
-
-<h3>Sign in (Multiple failed sign in attempts)</h3>
-<a href="sign-in">Sign in</a> =>
-<a href="sign-in?error=failed">Failed attempt</a> =>
-<a href="sign-in/troubleshoot">Trouble</a> =>
-<a href="sign-in/?error=last-attempt">Last attempt</a> =>
-<a href="account-locked">Account locked</a>
-<br />
-<br />
-
-<h3>Forgot password</h3>
-<a href="forgot-password">Forgot password</a> =>
-<a href="forgot-password?error=failed">Invalid email</a> =>
-<a href="forgot-password/check-email/">Check email</a>
-<br />
-<br />
-
-<h3>Reset password</h3>
-<a href="reset-password" target="_blank">Reset password</a><br />
-<a href="reset-password?error=password-mismatch">Password mismatch</a><br />
-<a href="reset-password?error=stronger-password-required">Password not strong enough</a><br />
-<a href="reset-password?error=password-too-common">Password too common</a><br />
-<a href="reset-password/confirmation">Confirmation</a><br />
-<br />
-
-<h3>Register</h3>
-<a href="register" target="_blank">Registration (flat happy path)</a><br />
