+++ app/templates/error.html
@@ -10,7 +10,7 @@
         <div class="panel panel--simple panel--error">
 
             <p>You have been signed out.</p>
-            <p><a href="{{ url_for('login') }}">Please Login.</a></p>
+            <p><a href="{{ url_for('sign_in_bp.login') }}">Please Login.</a></p>
 
         </div>
 {% endblock main %}
