+++ app/templates/login.html
@@ -11,7 +11,7 @@
     <div class="top-pad">
         <form
             method="POST"
-            action="{{ url_for('login') }}"
+            action="{{ url_for('sign_in_bp.login') }}"
             role="form">
 
             {{ form.csrf_token}}
