+++ app/templates/passwords/forgot-password.html
@@ -19,7 +19,7 @@
 <br />
 {% endif %}
 
-<form action="/forgot-password/check-email" class="form">
+<form action="/passwords/forgot-password/check-email" class="form">
 
     <h1 class="saturn">Forgot password</h1>
     <legend>No problem. We'll send you a link to reset your password.</legend>
