+++ app/templates/passwords/reset-password.html
@@ -38,7 +38,7 @@
 <br />
 {% endif %}
 
-<form action="/reset-password/confirmation" class="form">
+<form action="/passwords/reset-password/confirmation" class="form">
 
     <h1 class="saturn">Reset your password</h1>
     <p>Your password must have:</p>
@@ -71,7 +71,7 @@
 
     <button class="btn btn--primary venus" type="submit">Confirm</button>
 
-    <p><a href="{{ url_for('login') }}">Cancel</a></p>
+    <p><a href="{{ url_for('sign_in_bp.login') }}">Cancel</a></p>
 
 </form>
 {% endblock main %}
