+++ app/templates/register/register.almost-done.html
@@ -16,7 +16,7 @@
             <div class="details__content mars">
                 <div>
                     <p>Email not arrived? It might be in your spam folder.</p>
-                    <p>If it doesn’t arrive in the next 15 minutes, you can <a href="email-sent-again.html">request another email</a>.
+                    <p>If it doesn’t arrive in the next 15 minutes, you can <a href="/create-account/email-resent?user_id=TODO">request another email</a>.
                     </p></div>
             </div>
         </div>
