+++ app/templates/register/register.confirm-organisation-survey.html
@@ -7,7 +7,7 @@
 
 <form
         method="post"
-        action="{{ url_for('register_confirm_organisation_survey') }}"
+        action="{{ url_for('.register_confirm_organisation_survey') }}"
         class="form"
         role="form">
