+++ app/templates/register/register.email-resent.html
@@ -7,6 +7,6 @@
 
 <h1 class="saturn">We've re-sent your email</h1>
 
-{% include "register.help-with-email.html" %}
+{% include "register/register.help-with-email.html" %}
 
 {% endblock main %}
