+++ app/templates/register/register.enter-enrolment-code.html
@@ -23,7 +23,7 @@
 
     <form
             method="post"
-            action="{{ url_for('register') }}"
+            action="{{ url_for('.register') }}"
             class="form"
             role="form">
