+++ app/templates/register/register.enter-your-details.html
@@ -29,7 +29,7 @@
 
         <form
             method="post"
-            action="{{ url_for('register_enter_your_details') }}"
+            action="{{ url_for('.register_enter_your_details') }}"
             class="form"
             role="form">
