+++ app/templates/register/register.help-with-email.html
@@ -9,7 +9,7 @@
     <div class="details__main js-details-body" id="guidance-help-with-email" aria-hidden="true">
         <div class="details__content mars">
             <p>Email not arrived? It might be in your spam folder.</p>
-            <p>If it doesn’t arrive in the next 15 minutes, you can <a href="email-sent-again.html">request another
+            <p>If it doesn’t arrive in the next 15 minutes, you can <a href="/create-account/email-resent?user_id=TODO">request another
                 email</a>.</p>
         </div>
     </div>
