+++ app/templates/sign-in/sign-in.html
@@ -27,7 +27,7 @@
     <form
           method="post"
           class="form"
-          action="{{ url_for('login') }}"
+          action="{{ url_for('sign_in_bp.login') }}"
           role="form">
 
         {% if data['account_activated'] %}
@@ -35,7 +35,7 @@
             <p>You may now sign in.</p>
         {% else %}
             <p>Have an enrolment code?<br/>
-            Use it to <a href="/create-account">create an account</a>
+            Use it to <a href="/register/create-account">create an account</a>
             </p>
             <h1 class="saturn">Sign in</h1>
         {% endif %}
@@ -60,7 +60,7 @@
                     {{ form.password(class_='input input--text input-type__input js-password-obfuscation-field') }}
 
                     <br />
-                    <a href="/forgot-password">Forgot password?</a>
+                    <a href="/passwords/forgot-password">Forgot password?</a>
                 </div>
                 <br />
