+++ app/templates/sign-in/sign-in.last-attempt.html
@@ -8,7 +8,7 @@
     <form method="get" class="form" action="/sign-in/account-locked">
 
         <h1 class="saturn">Accessing your account</h1>
-        <p>If you're unsure of your sign in details, you may want to <a href="/forgot-password">reset your password.</a></p>
+        <p>If you're unsure of your sign in details, you may want to <a href="/passwords/forgot-password">reset your password.</a></p>
 
         <p>If you know your password, you can sign in below.</p>
