+++ app/templates/surveys/surveys-access.html
@@ -17,7 +17,8 @@
 <h2 class="neptune">Download</h2>
 
 
-<p>{{ survey_abbr }} spreadsheet for <strong>{{ business }}</strong> for the period {{ period_start }} to {{ period_end }}</p>
+<p>{{ survey_abbr }} spreadsheet for <strong>{{ business }}</strong> for the period {{ period_start }} to {{ period_end
+    }}</p>
 
 <p>If actual figures are not available, please provide informed estimates.</p>
 
@@ -40,36 +41,37 @@
 <p>Please upload by <strong>{{ submit_by }}</strong></p>
 
 <div class="guidance js-details upload-guidance"
-    data-guidance-label="Guidance for uploading"
-    data-guidance="Guidance for uploading"
-    data-hide-label="Hide guidance for uploading"
-    data-show-label="Show guidance for uploading">
+     data-guidance-label="Guidance for uploading"
+     data-guidance="Guidance for uploading"
+     data-hide-label="Hide guidance for uploading"
+     data-show-label="Show guidance for uploading">
 
-<a class="guidance__link js-details-trigger js-details-label mars"
-    id="guidance-for-uploading"
-    data-guidance-trigger="true"
-    aria-expanded="false"
-    href="#guidance-for-uploading"
-    aria-controls="guidance-help-with-email-body">Show guidance for uploading</a>
+    <a class="guidance__link js-details-trigger js-details-label mars"
+       id="guidance-for-uploading"
+       data-guidance-trigger="true"
+       aria-expanded="false"
+       href="#guidance-for-uploading"
+       aria-controls="guidance-help-with-email-body">Show guidance for uploading</a>
 
-<div class="guidance__main js-details-body" id="guidance-help-with-email" aria-hidden="true">
-    <div class="guidance__content mars">
-        <div>
-            <p>Your file needs to be:</p>
-            <ul>
-                <li>.xls or .xlsx format</li>
-                <li>smaller than 20MB</li>
-            </ul>
+    <div class="guidance__main js-details-body" id="guidance-help-with-email" aria-hidden="true">
+        <div class="guidance__content mars">
+            <div>
+                <p>Your file needs to be:</p>
+                <ul>
+                    <li>.xls or .xlsx format</li>
+                    <li>smaller than 20MB</li>
+                </ul>
+            </div>
         </div>
     </div>
-</div>
 
+</div>
 <br/>
 
 <p class="venus"><i class="fa fa-lock" aria-hidden="true"></i>&nbsp;&nbsp;&nbsp;
     We will treat your data securely and confidentially</p>
 
-{% include "surveys-upload-file-picker.html" %}
+{% include "surveys/surveys-upload-file-picker.html" %}
 
 
 {% endblock main %}
