+++ app/templates/surveys/surveys-history.html
@@ -8,13 +8,13 @@
 
 <div class="grid">
     <ul class="navigation-tabs navigation-tabs-component">
-        <li class="navigation-tabs__item navigation-tabs__item--primary navigation-tabs__item--active">
-            <a id="SURVEY_TODO_TAB" href="{{ url_for('surveys_bp.logged_in') }}"
-               class="btn btn--secondary btn--border navigation-tabs__tab navigation-tabs__tab--active">To do</a>
-        </li>
         <li class="navigation-tabs__item navigation-tabs__item--primary">
+            <a id="SURVEY_TODO_TAB" href="{{ url_for('surveys_bp.logged_in') }}"
+               class="btn btn--secondary btn--border navigation-tabs__tab">To do</a>
+        </li>
+        <li class="navigation-tabs__item navigation-tabs__item--primary navigation-tabs__item--active">
             <a id="SURVEY_HISTORY_TAB" href="{{ url_for('surveys_bp.surveys_history') }}"
-               class="btn btn--secondary btn--border navigation-tabs__tab">History</a>
+               class="btn btn--secondary btn--border navigation-tabs__tab navigation-tabs__tab--active">History</a>
         </li>
         <li class="navigation-tabs__item navigation-tabs__item--primary">
             <a id="SURVEY_MESSAGES_TAB" href="{{ url_for('secure_message_bp.messages_get') }}"
@@ -23,6 +23,6 @@
     </ul>
 </div>
 
-{% include "surveys.html" %}
+{% include "surveys/surveys.html" %}
 
 {% endblock main %}
