+++ app/templates/surveys/surveys-upload-failure.html
@@ -25,7 +25,7 @@
 
                 <div class="upload--response">
 
-                    {% include "surveys-upload-file-picker.html" %}
+                    {% include "surveys/surveys-upload-file-picker.html" %}
 
                 </div>
             </div>
