+++ app/views/surveys.py
@@ -7,7 +7,7 @@ from ons_ras_common.ons_decorators import jwt_session
 from app.config import Config
 
 logger = wrap_logger(logging.getLogger(__name__))
-surveys_bp = Blueprint('surveys_bp', __name__, static_folder='static', template_folder='templates')
+surveys_bp = Blueprint('surveys_bp', __name__, static_folder='static', template_folder='templates/surveys')
 
 
 def build_survey_data(status_filter):
@@ -40,7 +40,7 @@ def logged_in(session):
     # Get the survey data (To Do survey type)
     data_array = build_survey_data(status_filter)
 
-    return render_template('surveys-todo.html', _theme='default', data_array=data_array)
+    return render_template('surveys/surveys-todo.html', _theme='default', data_array=data_array)
 
 
 # ===== Surveys History =====
@@ -56,7 +56,7 @@ def surveys_history(session):
     data_array = build_survey_data(status_filter)
 
     # Render the template
-    return render_template('surveys-history.html',  _theme='default', data_array=data_array)
+    return render_template('surveys/surveys-history.html',  _theme='default', data_array=data_array)
 
 
 @surveys_bp.route('/access_survey', methods=['GET', 'POST'])
@@ -82,7 +82,7 @@ def access_survey(session):
     ci_data = req.json()
 
     # Render the template
-    return render_template('surveys-access.html', _theme='default', case_id=case_id, ci_data=ci_data,
+    return render_template('surveys/surveys-access.html', _theme='default', case_id=case_id, ci_data=ci_data,
                            survey=survey, survey_abbr=survey_abbr, business=business, period_start=period_start,
                            period_end=period_end, submit_by=submit_by)
 
@@ -113,15 +113,15 @@ def upload_survey(session):
 
     if result.status_code == 200:
         logger.debug('Upload successful')
-        return render_template('surveys-upload-success.html', _theme='default', upload_filename=upload_filename)
+        return render_template('surveys/surveys-upload-success.html', _theme='default', upload_filename=upload_filename)
     else:
         logger.debug('Upload failed')
         error_info = json.loads(result.text)
-        return render_template('surveys-upload-failure.html',  _theme='default', error_info=error_info,
+        return render_template('surveys/surveys-upload-failure.html',  _theme='default', error_info=error_info,
                                case_id=case_id)
 
 
 @surveys_bp.route('/surveys-upload-failure', methods=['GET'])
 def surveys_upload_failure():
     error_info = request.args.get('error_info', None)
-    return render_template('surveys-upload-failure.html', _theme='default', error_info=error_info)
+    return render_template('surveys/surveys-upload-failure.html', _theme='default', error_info=error_info)
